image-flash: explicitly use (long long) when printing off_t
The size of off_t changes with --{enable,disable}-largefile, so cast it to
(long long) to make sure printf always gets the correct value

+++ image-flash.c
@@ -51,8 +51,8 @@ static int flash_generate(struct image *image)
 			return -errno;
 
 		if (s.st_size > part->size) {
-			image_error(image, "image file %s for partition %s is bigger than partition (%ld > %lld)\n",
-					child->file, part->name, s.st_size, part->size);
+			image_error(image, "image file %s for partition %s is bigger than partition (%lld > %lld)\n",
+					child->file, part->name, (long long)s.st_size, part->size);
 			return -EINVAL;
 		}
 	}
